Improve structure and documentation of FilterClear connector
