+++ Sources/InstantSearchCore/FilterClear/Connector/FilterClearConnector.swift
@@ -11,7 +11,7 @@ import Foundation
 /// Component that clears all refinements that are currently active within the given FilterState
 public class FilterClearConnector {
   
-  /// Filter state that will hold your filters
+  /// Filter state holding your filters
   public let filterState: FilterState
   
   /// Logic applied to the clear control
@@ -25,7 +25,7 @@ public class FilterClearConnector {
   
   /**
    - Parameters:
-     - filterState: Filter state that will hold your filters
+     - filterState: Filter state holding your filters
      - interactor: Logic applied to the Clear Refinements
      - clearMode: Whether we should clear the specified filters or all filters except them
      - filterGroupIDs: The groupIDs of filters to clear. All filters will be cleared if unspecified.
@@ -44,50 +44,16 @@ public class FilterClearConnector {
   
 }
 
-extension FilterClearConnector {
-  
-  /**
-   - Parameters:
-     - filterState: Filter state that will hold your filters
-     - interactor: Logic applied to the Clear Refinements
-     - clearMode: Whether we should clear the specified filters or all filters except them
-     - filterGroupIDs: GroupIDs of filters to clear. All filters will be cleared if unspecified.
-     - controller: Controller that interfaces with a concrete clear refinement view
-   */
-  public convenience init(filterState: FilterState,
-                          interactor: FilterClearInteractor = .init(),
-                          clearMode: ClearMode = .specified,
-                          filterGroupIDs: [FilterGroup.ID]? = nil,
-                          controller: FilterClearController) {
-    self.init(filterState: filterState,
-              interactor: interactor,
-              clearMode: clearMode,
-              filterGroupIDs: filterGroupIDs)
-    connectController(controller)
-  }
-
-  /**
-   Establishes a connection with the controller
-   - Parameters:
-     - controller: Controller that interfaces with a concrete clear refinement view
-   - Returns: Established connection
-  */
-  @discardableResult func connectController(_ controller: FilterClearController) -> some Connection {
-    let connection = interactor.connectController(controller)
-    controllerConnections.append(connection)
-    return connection
-  }
-
-}
-
 extension FilterClearConnector: Connection {
   
   public func connect() {
     filterStateConnection.connect()
+    controllerConnections.forEach { $0.connect() }
   }
   
   public func disconnect() {
     filterStateConnection.disconnect()
+    controllerConnections.forEach { $0.disconnect() }
   }
   
 }
